Access the test cases from setup.py file
